Correction for CI errors

+++ tests/test_embedding_models.py
@@ -18,7 +18,8 @@ class TestEmbeddingsModels:
     def setup_class(cls):
         """Creation of a database and loads all the Embeddings Models. """
         cls.database_path = Path(f'cord19_{VERSION}.db')
-        os.remove(str(cls.database_path))
+        if cls.database_path.exists():
+            os.remove(str(cls.database_path))
         db = DatabaseCreation(data_path=Path('tests/data/'),
                               cord_path=Path('tests/data/CORD19_samples/'),
                               version=VERSION)
