+++ tests/test_sql.py
@@ -16,7 +16,8 @@ class TestDatabaseCreation:
     @classmethod
     def setup_class(cls):
         cls.database_path = Path(f'cord19_{VERSION}.db')
-        os.remove(str(cls.database_path))
+        if cls.database_path.exists():
+            os.remove(str(cls.database_path))
         db = DatabaseCreation(data_path=Path('tests/data/'),
                               cord_path=Path('tests/data/CORD19_samples/'),
                               version=VERSION)
